# Add code to switch to BitVec data structure in CSE optimization.