This change will allow us to track more than 64 CSE candidates.
Initially the data structure is set to the same limits (64) as is
currently used.  This gives us a no diffs.  Subsequent changes
will change these limits based on opportunities found in the benchmarks.

This change follows the same conventions that were used to move
assertion prop to use the BitVec type.

Code Diffs:
None - diffs were run using jit-diff on Ubuntu16.04 and Windows.  Additionally System.Private.CoreLib was dumped in a verbose mode that dumped the internal state of the bit vectors used in CSE.  The new data structure is bit for bit identical with the old data structure.

Throughput:
Timing of crossgen were taken on a number of assemblies.  The results were within the usual noise.
 |**Baseline**| |**Experiment**| |**Ex/Base**
:-----:|:-----:|:-----:|:-----:|:-----:|:-----:
Benchmark|Time(avg)|Score|Time(avg)|Score|Time
mscorlib-amd64|15539|0.4605|15509|0.4614|0.998069374
System.Data.Entity|12026|0.595|11991.2|0.5968|0.99710627
System.Windows.Forms|11564.4|0.6188|11646.6|0.6144|1.007108021
System.Design|7283.4|0.9825|7409.6|0.9658|1.017327073
System.Web-amd64|11453.6|0.6248|11621.8|0.6157|1.014685339
System.ServiceModel|18343.2|0.3901|18797.6|0.3807|1.024772123
Geomean|12205.89907|0.586261092|12325.44745|0.58058544|1.009794312