@russellhadley 2016-09-24
@briansull, @dotnet/jit-contrib PTAL